refator: Home 화면의 기존 내용을 ImagesContainer Componenet로 만들어서 관리하도록 변경 (#31)
