+++ src/components/MainImageContainer.jsx
@@ -1,0 +1,7 @@
+export default function MainImageContainder() {
+    return (
+        <>
+            <img src={"src/images/hotel_main.jpg"} alt="home_main" width="100%" height="100%" />
+        </>
+    )
+}
